@types Symlink + ParseCallback Overloading
Not sure how to use a similar thing to discriminated unions, but for function parameter types? I think this is the right way, one of these two, maybe with an interface, not totally sure. I was using it like this, and it wasn't working:

```ts
const result = await new Promise<RootTag>((resolve,reject) => parse(data,(error,result) => error !== null ? reject(error) : resolve(result)));
```

...but the result type doesn't narrow like I thought it would, I still just have to flip the conditional around to check if the result is null.

Oh yeah, and I was using this as-is for the type checking against the standard `nbt` package, right locally. With that new trick I learned the other day, I could install this one as an `@types/nbt` alias, `npm install -D @types/nbt@../nbt-js` (I have a demo folder adjacent to this one, so it's a symlinked-aliased installation). This added types for the package right away, and somehow I didn't have to add the `declare module "nbt"` syntax yet either. Not sure why that's the case.

And actually, I'm using this with the build of this code, not the original project. I still have to do that. So, one install in the demo folder, for this build (no types linked in the package.json), and the other aliased `@types/nbt` symlink.