Visibility Modifiers
Marking the properties as non-readonly again because I want the visibility state for things to represent the real scope used by the library itself. While these methods and properties likely should be privated because of their scoped use, that's not up to the types to make them fully hidden. If people have expected to be able to access them because you in fact can access them, then I want to ensure they can continue using things the same way as before, without any type errors. These are to guide you how you can use the library, not to ensure you use it a specific way.